code/src/api/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
eml_parser